ExOpt: optimisation-based exoskeleton control
This is part of a significant update to the 'APO-inverse-model' which
aims to achieve a number of things.
1) Generality and scalability to exoskeletons other than the APO;
2) A class-based rather than script-based coding structure which is
modular and scalable (this helps with (1))
3) Development of ExOpt as a framework which eventually can be used with
parts of other (existing or to-be-developed) Exoskeleton controllers
(formulated as force models).

This is a work in progress. Compared to the automatedInverseModel script
in Master, the code has been redesigned up to and including interfacing
with the OpenSim Matlab interface via the class OpenSimTrial. Future
work includes:
- Modification of getJointSpaceForces and introduction of
getExoskeletonJacobians
- Class structure for Optimisation and the overall control problem
- Structural changes and documentation