getJointSpaceForces and getFrameJacobians
Splitting getJointSpaceForces function in to two: getJointSpaceForces
(which only calculates the joint-space forces) and getFrameJacobians
(which gets the Jacobians - need these for the mapping of Exoskeleton
Forces on to the dynamic model).

getFrameJacobians reads in the points for which Jacobians should be
calculated using an xml settings file. These files are located in
Defaults/ContactPointSettings. The meat of this update is designing this
settings file (there's an example one for the apo), and also
implementing a parser using rapidxml for getFrameJacobians to interpret
these files.