Changes to getJointSpaceForces and getFrameJacobians
- They now produce time-indexed files. The reason they didn't was
previously for compatability with the optimisation script, but I think
it's better to handle this in Matlab. Should be easier to handle anyway,
what with the class-based structure.

-Added a check in getJointSpaceForces to ensure that the input files are
all time-aligned. Could do this before in Matlab, have two checks or
just one. Will have to see how it does... it just compares adding up the
times to doing 4*one of the times, so it might fail if the times are out
by even a little bit. About to do some testing so will see how it
performs.