JointSpaceForces
Began adding JointSpaceForces class to perform getJointSpaceForces and
store the results. However, given the extra complexities of subsampling
etc, I'm going to replace getJointSpaceForces with just doing ID on an
RRA result for the time being. I'll revisit this later and then I can
compare to see which is better.

Also made a small change to RRAResults to save the start and final time
corresponding to a RRA trial.