Testing getJointSpaceForces
Testing getJointSpaceForces, but can't finish fully as I don't want to
devote time to figuring out subsampling etc at the moment, also removing
header files etc. But I'll come back to this.

However, this did flag up an issue whereby using the OpenSimTrial
class's ID method doesn't result in a 1000Hz output file. This is
problematic and, from testing in the GUI, not expected. So I'll have to
deal with this.