Found why OpenSim ID tool was producing <1khz files
Thankfully this wasn't an issue with the code. Basically because I was
using the raw files output from RRA they still had the intermediate
timesteps in and this was causing OpenSim to think the framerate was
less than it actually was. When the data is fit to an RRAData object
these intermediate timesteps are removed. So this should be the process
i.e.

1) do RRA -> files generated
2) files => RRAData objects
3) RRAData objects -> files generated

Not sure whether to step 3 as a part of RRA or as and when necessary.
There's a danger of forgetting this step. I'll have a think on this and
create a post-it.