FrameJacobianSet, MatrixData, FrameJacobian
Rebranding of FrameJacobians class to FrameJacobianSet. It will likely
be made up of individual FrameJacobian objects which I will be making
later today.

This required the creation of a MatrixData class which I've implemented
and tested this morning. Basically reads in a data file with a sparse
time column where the remainder of the data corresponds to a sequence of
Matrices. Represents the data by giving a row and column size for the
matrices (assumed constant i.e. no variable-sized matrices within the
array), a frequency, a vector corresponding to a timesteps, and a cell
array each element of which is a matrix.