Change to RRAResults
A small amount of work on FrameJacobianSet and FrameJacobian. However,
mainly, since I'm about to be implementing the wrapping between Matlab
and getFrameJacobians, I wanted to be sure that I was supplying the
correct RRA files.

I'd run in to problems before because I was using the RRA files output
by OpenSim - not taking in to account the removeIntermediateTimesteps
feature of the RRAData class, which is important.

Therefore, I've now made it that when an RRAResults object is created it
reads in the OpenSim files, makes RRAData objects accordingly, then
reprints these to the same file destinations. Basically this removes the
intermediate timesteps from the files as well as the objects, making
them safe to use later.

Also I made a small change to OpenSimTrial so that time-ranges are
included in the output directory name - making it possible to run
several RRA's. However, given that an OST is associated with one
RRAResults object (which gets overwritten each time you run runRRA),
it's becoming clear to me that I think I need to separate OST and
RRAResults. Each RRAResult should be associated with one OST, but an OST
could be associated with many RRAResults. Basically an OST should only
be the constant things that don't change - model, kinematics, grfs. If
you changed any of these you'd really want a new OST, but if you just
wanted to run RRA from 0.5 to 1.0 instead of 0.5 to 0.6 you wouldn't
want to duplicate the OST object just to do this.

I'll be revisiting this later - but probably not for a while, as it's
relatively low down the priority list. I've added a post it for it.