Desired class, change to OfflineExOptController
Changing OfflineExOptController to separate out the results of a run
with the actual setting up of the parameters. So could run many runs
with different say optimisation schemes. This is similar to what I
eventually want to do with OpenSimTrial.

Also, working on implementing the Desired class. A mode will be given in
along with a varargin. Nothing can be done until we have the ID of the
RRA_q. But when we do, we evaluate the desired, and then we are ready to
perform the optimisation.