Additions to Data class
These changes allow to get the vector corresponding to a column headed
by a certain label, the index corresponding to a column headed by a
certain label.

Also allows to scale a column by some multiplier given the index and
multiplier. Have tested this and seems to work fine. Has a check to make
sure you're not scaling the time column.