Finished implementing percentage_reduction
Finished implementing the percentage_reduction methods of the Desired
class and tested these. Leaving the match_input methods for later. I
have all the inputs I need for the controller, now, so I can test it
against the results from the previous implementation.

Of course - I still have to update the optimisation function.