Exoskeleton class, separating OpenSimTrial from results
Begun implementing an Exoskeleton class. The idea here is that this will
hold relevant information pertaining to a particular exoskeleton e.g.
the contact point settings file, joint limits, a unique identifier
(name), etc. Not quite finished yet.

Also, began overhauling (slightly) the OpenSimTrial class to separate it
from the results. See commit 2f91863 for more discussion on the reasons
for this, but basically an OST should be defined on the constant
attributes like model file, kinematics, grfs - but not tied to a
particular RRA result since you may want to run a variety of RRA runs
with different parameters (start/end times, with adjustment or not,
etc). So, have result classes (RRAResults, IDResults, will add more at
some point probably). These are tied to an OpenSimTrial, but
OpenSimTrial objects are not tied to specific results.

I've made changes to the OST class to achieve this and included a new
IDResult class. Next, I need to go through the code and update the
dependences on the OpenSimTrial class where necessary.

There were also some other changes:

- Bug fixed in Data class within the averaged frequency calculation
(this was causing 50 timesteps at 100Hz to be interpreted as 98Hz)
- The optimisation classes were moved in to a separate folder since
they're a long term goal and I don't see myself working on them in the
near future.
- Removed the methods in OpenSimTrial to do RRA/ID from file. These
weren't actually working - the input directory was wrong - and anyway,
you would just use the default OpenSim Matlab interface for this if you
wanted to do this.