ExOpt-wide updates
Changes to some classes that had dependencies on the old OpenSimTrial to
still work with the updated version. Also, shifted some of the
functionality from the LinearExoskeletonForceModel class over to the
Exoskeleton class - Exoskeleton class now has the ability to create
predefined LinearExoskeletonForceModels. This feels more natural.

I haven't actually tested these changes, but I'm about to do so now, so
expect another commit with bugfixes/a confirmation that things are still
working as they should be.