ContactForceSet, adding functionality to LinearExoForceModel
- Implemented a ContactForceSet which stores spatial forces. Also, made
it so that the LinearExoskeletonForceModel has access to the
FrameJacobianSet and ContactForceSet calculated during its construction.

- Adding functionality to linearExoskeletonForceModel. Since it has
access to a ContactForceSet and FrameJacobianSet now, it can calculate
the spatial forces at the contact points given known trajectories (for
example). For the moment I've added support only for providing some
sample trajectories, which are stretched to the right size and used as
the exoskeleton motor torques in order to calculate the spatial forces.
This method could probably do with some work and generality though - it
seems like the ContactForceSet should also have knowledge of which motor
is 'linked' to each contact point.