Debugging and testing sptial force calculation from trajectory
This seems to be working. So, you can give a nominal motor torque
trajectory to an ExoskeletonForceModel, and it will compute the spatial
forces which are applied on the exoskeleton contact points due to these
motor torques.

This is not really useful for the optimisation, inverse direction type
thing. But, is more useful for running forward simulations (i.e. testing
in OpenSim) to analyse the accuracy of force models, perform some
validations, etc.