Small change to OpenSimTrial, more Data functionality
- Changes OpenSimTrial so that the load type must be specified in
construction. It makes no sense to not have this be the case, since the
grf data that you supply when constructing the trial is relative to a
certain loadType anyway.

- Added the ability to fit Data objects to a spline, and to combine data
objects (overloading addition) if they share the same timesteps. I'm
going to use this to to run forward simulations in OpenSim with APO
force model data. Basically I needed a way to create the updated grf
files which I now have almost completed.