Add functionality to LinearExoskeletonForceModel
This update mostly allows a LinearExoskeletonForceModel to take a
SpatialForceSet and write a new 'grf.mot' type file which describes the
original ground reaction forces and the reaction forces resulting from
the SpatialForceSet. This can then be used in an OpenSim simulation.
There were also a couple bug fixes, and the new routines were tested.

Together with the previous update to calculate a SpatialForceSet from a
trajectory, this allows the following:
1) Specify some motor torque trajectory;
2) Calculate the resulting spatial forces applied by an exoskeleton;
3) Using OpenSim and analyse the motion of the human-exoskeleton system
under the action of the specified motor trajectory.

This will allow 2 things: testing of how well the APO force model
performs (i.e. comparing transparent vs. active data with the full
forces applied in simulation), and also validation of the optimisation
scheme (use OpenSim to get torques from transparent and also with
assistance, try to get the optimisation to return the initial motor
torque trajectories).