Finish validating optimisation
Had to make a slight change to the optimisation validation procedure.
This is explained further in the comments for validateOptimisation, but
basically we only calculate RRA once and use the RRA kinematics to
compute both the grf-only ID and the grf + APO ID. This is as opposed to
doing RRA separately for each case.

Very close agreement between the actual and from-optimisation motor
torques now. Still should really double check that there's no accidental
timeshift anywhere i.e. in the Jacobians or something.