Modification to Desired Class
- Adding functionality to the Desired class so that it can match a
desired set of trajectories input as an IDResult. This still isn't
finished. As part of this, the function shift was added to the Data
class, which shifts one trajectory to match another based on the cross
correlation in the x direction. Basically, this is to account for the
fact that IDTrials will likely be input with some sort of lag or shift
between signals - this is to sort this out. Unfortunately it's not
possible to fully sort it out and we end up losing information at one of
the ends. I'll come back to fix this at some point.
- Also: put in place a new APO ContactPointSettings file - this change
hasn't been realised yet but basically this is what the
ExoskeletonSettings file could look like in the end. Will likely to an
update of this sort, but not yet.