Implementing Optimisation class
- Implemented Optimisation class which performs the optimisation based
on the requires inputs. This took away a bunch of functionality from the
OfflineController class, but this still exists as it takes an earlier
set of inputs and processes them to give them to the Optimisation.
OptimisationResult class holds the results of an Optimisation.
- Small change to data class 'fit to spline' in order to sure that the
sample points are generated fully. This involved rounding to 4 decimal
places, which should be fine for working with our 1000Hz data (0.001s
timesteps), however this would perhaps need to be made more fine if
working with higher frequency data.
- OfflineController and OfflineControllerResults changed to take in to
account of the Optimisation class.
- Made a change to RRAData which stops the last row being deleted. See
comment.
- Updated validateOptimisation to include the new Optimisation class.
Also, fixed a bug where the start time was set to 0.0 even though the IK
only started at 0.1, which was causing issues with numbers of frames.
- Updated OfflineController testing scripts.

This brings the changes up to previous levels of functionality. The last
thing to do on this branch is to introduce and test the ability to
declare joints as unconstrained using the constraintCoefficientMatrix
property of the desired class.