Reprocessing
Had to reprocess data to reflect some changes in the filtering process which brought raw data to marker trajectories/GRF files. A few changes here to accommodate this.

The processData line of this branch should be finished with now. However, the branch will remain open incase Graham needs to make more changes to any of the calculateMetric scripts.