Deleted image files
Deleted image files which were cluttering up the scripts folder. In general a large restructuring is needed for the metric stuff because we were working to a tight schedule. Will try to do this as I'm writing new stuff.

New images can be created via scripts, but should be stored away from the source code, and locally.