Remove RRA default gitignore
I was barking up the wrong tree here. What I think I'm going to have to do is have the default files as well as 'temp' ones which are created during setup and not tracked. But, this is definitely an issue to tackle later on, and not within this branch.