Change data structure
This script was used to change the structure of the data resulting from the data processing script. Ideally the script will be changed at a later date to make this unnecessary, but for now rerunning the entire data processing step was undesirable.

See comments of convertToStruct for more info on changes.