Revised metrics implementations
- Added a couply functions to replace calculations elsewhere (accountForTreadmill, isolateStancePhase). This also means that the method for isolating stance is now unified.
- Finished revised metrics implementations.

What follows is including anova testing and plotting in to metric class, splitting metric calculation code in to calculation and preparation, and a rethink of how the structure is saved.

I think it makes sense to have data.metrics be a list of named metric objects.

Will make notes and start implementing this.