Add File property
This acts as to keep a record of the filename which was loaded in.
Instead of using writeToFile with a filename argument you can just use
it with no arguments to overwrite the existing file.

This entry is required when creating  a Data object from properties, &
is useful for then printing said Data object.